pyproject.toml
[build-system]
requires = ['flit_core >=3.4,<4']
build-backend = 'flit_core.buildapi'

[project]
name = 'e3response'
dynamic = ["version", "description"]
authors = [
    { name = 'Martin Uhrin', email = 'martin.uhrin.10@ucl.ac.uk' },
]
readme = 'README.rst'
license = { file = 'LICENSE.txt' }
classifiers = [
    'Development Status :: 4 - Beta',
    'License :: OSI Approved :: GNU Lesser General Public License v3 or later (LGPLv3+)',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
]
keywords = ['machine learning', 'e3nn-jax', 'physics']
requires-python = '>=3.9'
dependencies = [
    "ase",
    "hydra-core",
    'jax',
    'jaxlib',
    'jraph',
    'orbax-checkpoint',
    'e3nn-jax',
    "equinox",
    "reax>=0.2.0",
    "tensorial",
]

[project.urls]
Home = 'https://github.com/camml-lab/e3response'
Source = 'https://github.com/camml-lab/e3response'

[project.optional-dependencies]
docs = [
    'docutils',
    'jupyter', # For running doc examples
    'nbsphinx', # Jupyter notebooks in docs
    'pandoc',
    'sphinx',
    'sphinx-autobuild',
]
dev = [
    'flit',
    'ipython',
    'pytest',
    'pytest-cov',
    'pre-commit',
    'pylint',
    'black',
]
[project.scripts]
e3response = "e3response.cli:main_cli"

[tool.bandit]
exclude_dirs = ["test/"]
skips = ["B101"]

[tool.flit.module]
name = 'e3response'

[tool.flit.sdist]
exclude = [
    '.github/',
    'docs/',
    'examples/',
    'test/',
]

[tool.flynt]
line-length = 100
fail-on-change = true

[tool.isort]
profile = "black"
force_sort_within_sections = true
include_trailing_comma = true
line_length = 100
multi_line_output = 3

[tool.pylint.format]
max-line-length = 100

[tool.black]
line-length = 100

[tool.pylint.messages_control]
disable = [
    'duplicate-code',
    'import-outside-toplevel',
    'missing-docstring',
    'locally-disabled',
    'too-few-public-methods',
    'too-many-arguments',
    'too-many-instance-attributes',
    'use-dict-literal',
]

[tool.pytest.ini_options]
minversion = '6.0'
testpaths = [
    'test',
]
filterwarnings = [
    'ignore::DeprecationWarning:frozendict:',
]

[tool.yapf]
align_closing_bracket_with_visual_indent = true
based_on_style = 'google'
coalesce_brackets = true
column_limit = 100
dedent_closing_brackets = true
indent_dictionary_value = false
split_arguments_when_comma_terminated = true

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py311

[testenv]
usedevelop = true

[testenv:py{39,310,311,312}]
description = Run the unit tests
extras =
    dev
commands = pytest {posargs}

[testenv:pre-commit]
description = Run the style checks and formatting
extras =
    dev
commands = pre-commit run {posargs}

[pytest]
filterwarnings =
    ignore::DeprecationWarning:distutils:
"""
